recipe-suggester-backend/requirements.txt
# Django & Required Libraries
Django==5.0.1
djangorestframework==3.14.0
Pillow==10.2.0  # Required for handling images

# Database & ORM
sqlparse==0.4.4  # Required by Django for SQLite

# Authentication & Security
argon2-cffi==23.1.0  # For password hashing

# ASGI & WebSocket Support (if using Django Channels)
asgiref==3.7.2
channels==4.1.0

# Background Tasks (Celery)
celery==5.4.0
kombu==5.3.7
amqp==5.2.0
billiard==4.2.0
vine==5.1.0

# API Utilities
requests==2.32.3
urllib3==1.26.18
charset-normalizer==3.3.2
idna==3.6

# Optional Dependencies for Additional Features
pytz==2023.3.post1  # Timezone support
tzdata==2023.4  # Timezone data
six==1.16.0  # Required for compatibility in some libraries
